slowly

+++ ft_strncmp.c
@@ -6,16 +6,18 @@
 /*   By: mberquer <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/29 12:27:11 by mberquer          #+#    #+#             */
-/*   Updated: 2021/11/29 12:32:02 by mberquer         ###   ########.fr       */
+/*   Updated: 2021/11/30 17:04:21 by mberquer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 int	ft_strncmp(const char	*s1, const char	*s2, size_t	n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
-	while(s1[i] == s2[i] && i < n)
+	while (s1[i] == s2[i] && i < n)
 		i++;
 	return (s1[i] - s2[i]);
 }
